Only enable multicast rules when needed
In corosync 2.x multicast is the default, but if 'udpu' transport is
explicitly configured, we don't need to allow it.

Also, we filter incoming Multicast packets by their source.

Default behaviour prior to this patch series was to always allow
unicast, this has been kept (also needed to run current tests).

Signed-off-by: Stefan Reiter <s.reiter@proxmox.com>