helpers: add helper for checking firewall bridge creation
The logic for checking whether a firewall bridge should get created
was duplicated across several call sites in qemu-server and
pve-container. This makes updating the condition error-prone and might
introduce bugs due to different logic. Create a new helper that the
call sites can use instead of having to re-implement the logic there.

Previously, call sites creating a network device only checked whether
nftables was enabled or not. This commit adds a check whether the
bridge is a Linux bridge as well. This is required because we still
have to create firewall bridges for OVS bridges, since otherwise the
guest firewall won't work. Netfilter doesn't see traffic passing
through a OVS bridge, only through a normal Linux bridge. This causes
the guest firewall to not work on OVS bridges when nftables is
enabled.

The preferred place would be the tap_plug function, but since we need
to read the firewall configuration there this would create a circular
dependency between pve-firewall and pve-common, so we put the helper
here instead and call it in the tap_plug wrappers in qemu-server and
pve-container.

This commit moves the is_nftables helper to PVE::Firewall::Helper to
avoid a circular dependency between the Firewall and the Helper
module. It replaces the existing function with a call to the new
function in PVE::Firewall::Helper for backwards compatibility.

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>

FG: the helper in pve-network is not meant for general purpose usage, to
avoid entangling things just inline it here

Signed-off-by: Fabian Grünbichler <f.gruenbichler@proxmox.com>